fix(ci): file-length guard — decompose execution_flow + part_002 under 1500-line limit
PR #545's artifact work pushed two first-party files past the
MAX_FIRST_PARTY_FILE_LINES=1500 maintainability guard
(test_core_decomposition_maintainability), failing the coverage shard:
execution_flow.py (1549) and test_executor_part_002.py (1511).

Extract the planning-artifact concern from execution_flow.execute into a
new sibling module executor/planning_artifacts.py
(_deposit_planning_artifacts_best_effort + handle_agent_planning_result),
mirroring the existing execution_validation extraction pattern; behavior
is unchanged. Relocate the planning-validation cleanup-failure test from
test_executor_part_002.py to test_executor_part_005.py, which already
shares the same fixtures and seed helpers.
