+++ src/awf/control/executor/planning_artifacts.py
@@ -1,0 +1,108 @@
+"""WorkspaceExecutor planning-artifact handling.
+
+Mechanically extracted from ``execution_flow.execute``; behavior is unchanged.
+Keeping the plan/conformance-report deposit and the agent/planning result
+interpretation here keeps ``execution_flow`` focused on the agent-run →
+commit → validate → push pipeline.
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+from typing import Any
+
+from awf.control.executor.helpers import _failure_salvage_payload
+from awf.control.executor.types import _PlanningRunFailure, _PlanningValidationHandoff
+from awf.db.enums import FailureReason, WorkspaceStatus
+from awf.runtime.planning import AGENT_PLAN_PHASE_SCOPE_VIOLATION, render_workspace_path
+from awf.service.artifacts import deposit_workspace_planning_artifacts
+
+
+def _deposit_planning_artifacts_best_effort(
+    self: Any,
+    *,
+    profile: Any,
+    workspace_id: str,
+    worktree_path: Path,
+) -> None:
+    """Deposit the worktree plan + conformance report into the served artifact
+    dir (a sibling of the worktree) before any teardown removes the worktree,
+    so the console can surface them. Gated on the resolved planning profile —
+    the only reliable "planning was required" signal (the handoff is set only
+    on the AWF-validation path, not when conformance is satisfied inline).
+    Runs on the success path, the validation/conformance stop paths, and the
+    early planning-failure return — all preserve a FAILED workspace whose
+    worktree still holds the plan/report. A copy failure is best-effort and
+    non-fatal.
+    """
+    if profile is None or not profile.planning.required:
+        return
+    deposit_workspace_planning_artifacts(
+        work_dir=self._config.compose_projects_root.parent,
+        workspace_id=workspace_id,
+        worktree_path=worktree_path,
+        plan_path=render_workspace_path(profile.planning.plan_path, workspace_id=workspace_id),
+        report_path=render_workspace_path(
+            profile.planning.conformance_report_path, workspace_id=workspace_id
+        ),
+    )
+
+
+async def handle_agent_planning_result(
+    self: Any,
+    *,
+    workspace_id: str,
+    ws: Any,
+    worktree_path: Path,
+    profile: Any,
+    planning_failure: Any,
+) -> tuple[_PlanningValidationHandoff | None, bool]:
+    """Interpret the agent/planning run result.
+
+    Returns ``(handoff, should_return)`` where ``handoff`` is the planning
+    validation handoff to thread into validation (``None`` when none applies)
+    and ``should_return`` signals that ``execute`` must stop because the
+    workspace was already marked FAILED.
+    """
+    if isinstance(planning_failure, _PlanningValidationHandoff):
+        await self._record_planning_validation_handoff_event(
+            workspace_id=workspace_id,
+            handoff=planning_failure,
+        )
+        return planning_failure, False
+    if planning_failure is not None:
+        failure_message = (
+            planning_failure if isinstance(planning_failure, str) else planning_failure.message
+        )
+        reason_code = None if isinstance(planning_failure, str) else planning_failure.reason_code
+        details = None if isinstance(planning_failure, str) else planning_failure.details
+        await self._mark_failed(
+            workspace_id=workspace_id,
+            from_status=WorkspaceStatus.running,
+            failure_reason=FailureReason.agent_failure,
+            message=failure_message[:2000],
+            reason_code=reason_code,
+            details=details,
+            salvage=_failure_salvage_payload(ws, worktree_path=worktree_path),
+        )
+        # The planning loop preserves the FAILED workspace's worktree,
+        # which can already hold the plan + (unsatisfied) conformance
+        # report. Deposit them now — this branch returns before the
+        # post-validation deposit block, so without this the console
+        # could not surface the failed report.
+        _deposit_planning_artifacts_best_effort(
+            self,
+            profile=profile,
+            workspace_id=workspace_id,
+            worktree_path=worktree_path,
+        )
+        if (
+            isinstance(planning_failure, _PlanningRunFailure)
+            and planning_failure.reason_code == AGENT_PLAN_PHASE_SCOPE_VIOLATION
+        ):
+            await self._auto_retry_planning_scope_failure(
+                workspace_id=workspace_id,
+                failure=planning_failure,
+            )
+        return None, True
+    return None, False
